Print names of files before parsing
In the case where a file has invalid JSON data in it, it can be
difficult to debug because the tool previously didn't give any
indication of which file had the issue.  This commit just adds
some debug output that should make it easier to identify which
file may be malformed.